+++ package.json
@@ -13,7 +13,8 @@
 		"format": "prettier --write .",
 		"test:integration": "playwright test",
 		"test:unit": "vitest",
-		"deploy": "npm run build && gsutil -m cp -R build/* gs://caf-planning-staticfiles"
+		"deploy": "npm run build && gsutil -m cp -R build/* gs://caf-planning-staticfiles",
+		"psql": "PGPASSWORD=postgres psql -U bcoste --host 127.0.0.1 --port 5432 -d caf"
 	},
 	"devDependencies": {
 		"@floating-ui/dom": "^1.5.3",
