+++ src/routes/utils.ts
@@ -1,29 +1,39 @@
 import { DateTime } from 'luxon';
 
 export function create_slots() {
-    const slots = {};
+	const slots = {};
 
-    const today = DateTime.now().setZone('utc+0', { keepLocalTime: true }).startOf('day');
+	const today = DateTime.now().setZone('utc+0', { keepLocalTime: true }).startOf('day');
 
-    for (let i = 0; i < 20; i++) {
-        const day = today.plus({ days: i });
+	for (let i = 0; i < 20; i++) {
+		const day = today.plus({ days: i }) as DateTime;
 
-        let item = {
-            id: i,
-            day: day,
-            month: day.month,
-            date: day.toLocaleString({ weekday: 'long', day: 'numeric', month: 'long' }),
-            status: 'ok',
-            adding_slot: false,
-            refs: {}
-        };
+		if (![1, 2, 5].includes(day.weekday)) {
+			continue;
+		}
+		const item = {
+			id: i,
+			day: day.toISO().slice(0, 10),
+			month: day.month,
+			date: day.toLocaleString({ weekday: 'long', day: 'numeric', month: 'long' }),
+			status: 'ok',
+			adding_slot: false,
+			refs: {}
+		};
 
-        // let cancelled_day = new Date(2024, 0, 19);
-        // if (day.toDateString() == cancelled_day.toDateString()) {
-        //     item.status = "cancelled";
-        // }
+		// let cancelled_day = new Date(2024, 0, 19);
+		// if (day.toDateString() == cancelled_day.toDateString()) {
+		//     item.status = "cancelled";
+		// }
 
-        slots[day] = item;
-    }
-    return slots
+		slots[item.day] = item;
+	}
+	return slots;
 }
+
+export const timeslots = [];
+for (let i = 0; i < 4; i++) {
+	timeslots.push(`${18 + i}:00`);
+	timeslots.push(`${18 + i}:30`);
+}
+timeslots.push(`22:00`);
